[Context] Extract the test path from the coverage response.
We were wrongfuly assuming the path of the test was the original file, so if `my_feature.py` was related to `test.py::test1` and `test2.py::other_test`, we were masshing both together as `my_feature.py::test.py::test1 my_feature.py::test2.py::other_test`.

With this PR, we not only fix that, but also cover the case where a single surce file covers different test files.